+++ src/components/index.js
@@ -1,0 +1,7 @@
+export { default as createButton } from './Button.js';
+export { default as createHeader } from './Header.js';
+export { default as createInput } from './Input.js';
+export { default as createModal } from './Modal.js';
+export { default as createRestaurantItem } from './RestaurantItem.js';
+export { default as createSelect } from './Select.js';
+export { default as createTextarea } from './Textarea.js';
